Prevent URL trailing char errors in VIM/WIM
It is very common for users to copy/paste the URL in the web GUI form
when registering new VIMs/WIMs.
This can be error prone, since it might introduce trailing characters
that might pass unnoticed (e.g. when copying from "OpenStack Rocky >
Project > API Access" a trailing space will be introduced).

This change tries to prevent that from happening.

Change-Id: I8ef83afe39ef854696ca43072273437ffd396c33
Signed-off-by: Anderson Bravalheri <a.bravalheri@bristol.ac.uk>