Fix osm-common and RO-plugin dependencies to use v9.0
The dependencies in requirements.txt were pointing to master
instead of pointing to v9.0. With the recent changes in the
repos, osm-common and RO-plugin do not have dependencies,
so errors might appear if this change is not applied.

The file setup.py was also updated because it was enough to specify
the requirements there as 'RO-plugin' and 'osm-common'.

Change-Id: I56a922364899fdd73eeb2bf6210f89d9f9ed14d0
Signed-off-by: garciadeblas <gerardo.garciadeblas@telefonica.com>