Find the interface ports regardless of VM's existence
This patch will find and delete the network interface ports regardless of VM's existence.
In healing feature, when the VM is deleted manually, network interfaces ports still
exist in the Openstack Cloud. If we check the network ports' presence according to vm_id,
those ports are not deleted and heal operation fails. This patch will find all the ports,
than deletion required interface will be deleted.

Change-Id: I139559844aa148109588f70f1472ee2e07d7df67
Signed-off-by: aticig <gulsum.atici@canonical.com>