Fix persistent volume creation after latest changes in IM
This change fixes the issues with persistent volume creation.
After latest changes in IM, the IM generated from yang models changed
and type-of-storage was unexpectedly prefixed with the name of the YANG
module, leading to the RO code to undetect volume creation.

This change also adds some logs to improve the debugging in the future.

Change-Id: If942113a6e4feebaf2c2bc499f27de5d4439fa9d
Signed-off-by: garciadeblas <gerardo.garciadeblas@telefonica.com>